refactor: moved release substructs to their own folders
