+++ musicbrainz_db_lite/src/models/musicbrainz/mod.rs
@@ -1,3 +1,4 @@
+pub mod label_info;
 pub mod media;
 use std::sync::Arc;
