+++ musicbrainz_db_lite/src/models/musicbrainz/release/mod.rs
@@ -84,28 +84,6 @@ impl HasArtistCredits<MBRelease> for Release {
     }
 }
 
-
-
-#[derive(Debug, Default, Clone, FromRow, Upsert)]
-#[database(
-    table = "label_infos",
-    primary_key = "id",
-    ignore_insert_keys(id),
-    ignore_update_keys(id, gid)
-)]
-pub struct LabelInfo {
-    pub id: i64,
-    pub catalog_number: Option<String>,
-    pub label: Option<String>,
-    pub release: i64,
-}
-
-impl crate::RowId for LabelInfo {
-    fn get_row_id(&self) -> i64 {
-        self.id
-    }
-}
-
 impl HasTable for Release {
     const TABLE_NAME: &str = "releases";
     const FOREIGN_FIELD_NAME: &str = "release";
