+++ musicbrainz_db_lite/src/models/musicbrainz/release/relations/label_infos.rs
@@ -1,6 +1,6 @@
 use sqlx::SqliteConnection;
 
-use crate::models::musicbrainz::release::LabelInfo;
+use crate::LabelInfo;
 use crate::models::musicbrainz::release::Release;
 
 impl Release {
